+++ foundation-runtimes/terminal/api/src/commonMain/kotlin/terminal/Terminal.kt
@@ -1,3 +1,10 @@
 package terminal
 
-interface Terminal : TerminalInput, TerminalOutput
+interface Terminal : TerminalOutput {
+    fun executeReadLine(prompt: String): String
+    fun readln(prompt: String = ""): String {
+        val res = executeReadLine(prompt)
+        history.push(res)
+        return res
+    }
+}