+++ foundation-runtimes/terminal/api/src/commonMain/kotlin/terminal/TerminalInput.kt
@@ -1,5 +1,0 @@
-package terminal
-
-interface TerminalInput {
-    fun readln()
-}