+++ foundation-runtimes/terminal/console/src/commonMain/kotlin/terminal/ConsoleTerminalOutput.kt
@@ -1,6 +1,8 @@
 package terminal
 
-class ConsoleTerminalOutput(config: TerminalHistoryConfig = TerminalHistoryConfig()) : TerminalOutput {
+class ConsoleTerminalOutput(
+    config: TerminalHistoryConfig = TerminalHistoryConfig()
+) : TerminalOutput {
     override val history = TerminalHistory(config)
     override fun <T> executePrint(any: T) = kotlin.io.print(any)
 }